Refactor model_validation to api_utils
